任務功能:建立任務，改名任務，刪除任務

+++ src/main/java/com/kazama/springtodolistweb/Dao/impl/TaskDaoImpl.java
@@ -1,7 +1,10 @@
 package com.kazama.springtodolistweb.Dao.impl;
 
 import com.kazama.springtodolistweb.Dao.TaskDao;
+import com.kazama.springtodolistweb.RowMapper.TaskRowMapper;
 import com.kazama.springtodolistweb.dto.TaskRequest;
+import com.kazama.springtodolistweb.dto.UpdateTaskRequest;
+import com.kazama.springtodolistweb.model.Task;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
 import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
@@ -11,6 +14,7 @@ import org.springframework.stereotype.Component;
 
 import java.util.Date;
 import java.util.HashMap;
+import java.util.List;
 import java.util.Map;
 
 @Component
@@ -19,15 +23,17 @@ public class TaskDaoImpl implements TaskDao {
     @Autowired
     private NamedParameterJdbcTemplate namedParameterJdbcTemplate;
 
-    public Integer addTask(TaskRequest taskRequest){
+    public Integer addTask(Integer userId ,Integer listId ,TaskRequest taskRequest){
 
-        String sql = "INSERT INTO task(task_name , created_date , deadline) VALUES (:task_name ,:created_date, :deadline)";
+        String sql = "INSERT INTO task(task_name , created_date , deadline ,user_id,list_id) VALUES (:task_name ,:created_date, :deadline , :user_id ,:list_id) ";
 
         Map<String, Object> map = new HashMap<>();
 
         map.put("task_name" ,taskRequest.getTaskName());
         map.put("deadline", new Date());
         map.put("created_date", new Date());
+        map.put("user_id" , userId);
+        map.put("list_id",listId);
 
         KeyHolder keyHolder = new GeneratedKeyHolder();
         namedParameterJdbcTemplate.update(sql , new MapSqlParameterSource(map) ,keyHolder);
@@ -37,14 +43,47 @@ public class TaskDaoImpl implements TaskDao {
         return key ;
     }
 
-    public void deleteTaskById(Integer taskId){
-        String sql = "DELETE FROM task  WHERE task_id =:taskId";
+    public void deleteTask(Integer userId,Integer listId,Integer taskId){
+        String sql = "DELETE FROM task  WHERE task_id =:taskId AND user_id =:userId AND list_id =:listId";
 
         Map<String,Object> map = new HashMap<>();
 
         map.put("taskId", taskId);
+        map.put("userId", userId);
+        map.put("listId", listId);
+
 
         namedParameterJdbcTemplate.update(sql,map);
+
+    }
+
+    @Override
+    public Task getTaskById(Integer taskId) {
+        String sql = "SELECT task_id , task_name,created_date,deadLine,user_id,list_id FROM task WHERE task_id = :task_id";
+        Map<String ,Object> map  = new HashMap<>();
+
+        map.put("task_id" ,taskId);
+
+        List<Task> taskList = namedParameterJdbcTemplate.query(sql ,new MapSqlParameterSource(map) ,new TaskRowMapper());
+
+        if(taskList.size()>0) return taskList.get(0);
+        else return null;
+
+    }
+
+    @Override
+    public void renameTask(Integer userId, Integer listId, UpdateTaskRequest updateTaskRequest){
+        String sql = "UPDATE task SET task_name = :updateName WHERE user_id =:userId AND list_id =:listId AND task_id =:taskId";
+
+        Map<String, Object> map = new HashMap<>();
+
+        map.put("userId",userId);
+        map.put("listId",listId);
+        map.put("updateName",updateTaskRequest.getUpdateName());
+        map.put("taskId" , updateTaskRequest.getTaskId());
+
+
+        namedParameterJdbcTemplate.update(sql , map);
 
     }
 }
